Added the Identity Submenu, and did a lot of clean up/debugging
Reconfigured BlendShapeConfiguration.cs and created CharacterFile.cs such that characeter settings can be saved and loaded as JSON files. This could be useful for creating and loading NPCs or other characters.

These functions are accessible through the new Identity Submenu. In this submenu you can also name your character and reset your character back to the default settings.

Had to modify CharacterSlider.cs and BiologyMenu.cs such that the sliders accurately readjusted when a character file is loaded.

Added an image for the load button.

A ton of misc. cleaning up and error prevention.